requirements.txt
pandas>=1.5.0
numpy>=1.23.0
matplotlib>=3.6.0
plotly>=5.11.0
seaborn>=0.12.0
scikit-learn>=1.2.0
scipy>=1.9.0
pytest>=7.2.0
black>=22.12.0
flake8>=6.0.0
python-dotenv>=0.21.0
pyyaml>=6.0